requirements.txt
ipfml
sklearn
scikit-image
tensorflow
keras
image_slicer
Pillow
pydot
matplotlib
path.py
pandas
opencv-python
gzip